End of Section 18

+++ client/src/app/messages/messages.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { Message } from '../_models/message';
 import { Pagination } from '../_models/Pagination';
+import { ConfirmService } from '../_services/confirm.service';
 import { MessageService } from '../_services/message.service';
 
 @Component({
@@ -9,15 +10,15 @@ import { MessageService } from '../_services/message.service';
   styleUrls: ['./messages.component.css']
 })
 export class MessagesComponent implements OnInit {
-messages : Message[] = [];
-pagination : Pagination;
-container = 'Unread';
-pageNumer = 1;
-pageSize = 5;
-loading = false;
-messagePresent = false;
+  messages: Message[] = [];
+  pagination: Pagination;
+  container = 'Unread';
+  pageNumer = 1;
+  pageSize = 5;
+  loading = false;
+  messagePresent = false;
 
-  constructor(private messageService : MessageService) { }
+  constructor(private messageService: MessageService, private confirmService: ConfirmService) { }
 
   ngOnInit(): void {
     this.loadMessages();
@@ -26,20 +27,25 @@ messagePresent = false;
 
   loadMessages() {
     this.loading = true;
-    this.messageService.getMessages(this.pageNumer,this.pageSize,this.container).subscribe(response =>{
+    this.messageService.getMessages(this.pageNumer, this.pageSize, this.container).subscribe(response => {
       this.messages = response.result;
       this.pagination = response.pagination;
       this.loading = false;
     })
   }
 
-  deleteMessage(id : number){
-    this.messageService.deleteMessage(id).subscribe(()=>{
-      this.messages.splice(this.messages.findIndex(m => m.id == id), 1);
+  deleteMessage(id: number) {
+    this.confirmService.confirm('Confirm delete message', 'This cannot be undone').subscribe(result => {
+      if (result) {
+        this.messageService.deleteMessage(id).subscribe(() => {
+          this.messages.splice(this.messages.findIndex(m => m.id == id), 1);
+        })
+      }
     })
+
   }
 
-  pageChanged(event : any){
+  pageChanged(event: any) {
     this.pageNumer = event.page;
     this.loadMessages();
   }
